+++ app/src/main/java/sk/hello/movies/modules/SortingModule.java
@@ -2,22 +2,21 @@ package sk.hello.movies.modules;
 
 import dagger.Module;
 import dagger.Provides;
-import sk.hello.movies.interfacers.SortingDialogInteractorImpl;
-import sk.hello.movies.interfacers.SortingDialogPresenterImpl;
+import sk.hello.movies.interfacers.SortingDialogInterfacer;
+import sk.hello.movies.interfacers.SortingDialogPresenter;
 import sk.hello.movies.utils.SortingOptionStore;
 import sk.hello.movies.interfaces.SortingDialogInterface;
-import sk.hello.movies.interfaces.SortingDialogPresenter;
 
 
 @Module
 public class SortingModule {
     @Provides
     public SortingDialogInterface providesSortingDialogInteractor(SortingOptionStore store) {
-        return new SortingDialogInteractorImpl(store);
+        return new SortingDialogInterfacer(store);
     }
 
     @Provides
-    public SortingDialogPresenter providePresenter(SortingDialogInterface interactor) {
-        return new SortingDialogPresenterImpl(interactor);
+    public sk.hello.movies.interfaces.SortingDialogPresenter providePresenter(SortingDialogInterface interactor) {
+        return (sk.hello.movies.interfaces.SortingDialogPresenter) new SortingDialogPresenter(interactor);
     }
 }
